Move mpm88 runner into separate file
