+++ app/src/main/assets/mpm88_field_shaders/metadata.json
