+++ app/src/main/java/com/example/mpm88ndarray/MpmObstacleRunner.java
@@ -1,0 +1,4 @@
+package com.example.mpm88ndarray;
+
+public class MpmObstacle {
+}
